+++ hrswp-blocks.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: HRSWP Blocks
- * Version: 3.1.1
+ * Version: 3.2.0
  * Description: A WSU HRS WordPress plugin to provide custom blocks and WP block editor adjustments.
  * Author: Adam Turner, washingtonstateuniversity
  * Author URI: https://hrs.wsu.edu/
@@ -9,8 +9,8 @@
  * Update URI: https://api.github.com/repos/washingtonstateuniversity/hrswp-plugin-blocks/releases/latest
  * Text Domain: hrswp-blocks
  * Requires at least: 5.9
- * Tested up to: 6.0.0
- * Requires PHP: 7.3
+ * Tested up to: 6.0.3
+ * Requires PHP: 7.4
  *
  * @package HRSWP_Blocks
  * @since 0.1.0
@@ -38,14 +38,15 @@ register_activation_hook(
 	}
 );
 
-// Load blocks.
+// Load blocks, asset loader, settinsg, and REST API.
 require_once dirname( __FILE__ ) . '/inc/blocks.php';
-
-// Load the asset loader.
 require_once dirname( __FILE__ ) . '/inc/asset-loader.php';
-
-// Load the plugin settings.
 require_once dirname( __FILE__ ) . '/inc/settings.php';
+require_once dirname( __FILE__ ) . '/inc/api.php';
+require_once dirname( __FILE__ ) . '/inc/query.php';
+
+// Load plugin classes.
+require_once dirname( __FILE__ ) . '/inc/classes/class-sideload-image.php';
 
 /**
  * Uninstalls the plugin.
@@ -62,3 +63,35 @@ function uninstall(): void {
 	delete_option( 'hrswp_plugins_protected_ids' );
 }
 register_uninstall_hook( __FILE__, __NAMESPACE__ . '\uninstall' );
+
+/**
+ * Verifies plugin dependencies.
+ *
+ * @since 3.2.0
+ *
+ * @return bool True if dependencies are met, false if not.
+ */
+function verify_plugin_dependencies(): bool {
+	return in_array(
+		'hrswp-plugin-sqlsrv-db/hrswp-sqlsrv-db.php',
+		apply_filters( 'active_plugins', get_option( 'active_plugins' ) ),
+		true
+	);
+}
+
+add_action(
+	'plugins_loaded',
+	function(): void {
+		if ( ! verify_plugin_dependencies() ) {
+			add_action(
+				'admin_notices',
+				function(): void {
+					printf(
+						'<div class="error"><p>%s</p></div>',
+						esc_html__( 'The HRSWP Blocks plugin requires the HRSWP Sqlsrv DB plugin to function properly. Please install before continuing.', 'hrswp-blocks' )
+					);
+				}
+			);
+		}
+	}
+);