+++ src/blocks/filter-section/icon.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-const { Path, SVG } = wp.components;
+import { Path, SVG } from '@wordpress/components';
 
 const search = (
 	<SVG xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">