+++ src/blocks/posts-list/icons.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { SVG, Path } from '@wordpress/primitives';
+import { Path, SVG } from '@wordpress/components';
 
 export const icon = (
 	<SVG xmlns="http://www.w3.org/2000/svg" viewBox="-2 -2 24 24">