+++ src/store/defaults.js
@@ -1,0 +1,11 @@
+/**
+ * Default value for the job position data.
+ *
+ * @type {Object}
+ */
+export const POSITION_DATA_DEFAULTS = {
+	tableNames: {},
+	jobClassificationData: {},
+	salaryData: {},
+	awardsData: {},
+};